fixed: https://mootools.lighthouseapp.com/projects/24057/tickets/332-drag-webkit-issue
issue caused when element had a set right or bottom position. Assuming it was a set right position, then (in WebKit) the left would become 'auto,' and adjusting left position would fail.

So, I detection which triggers an update to the modifier and an invert of the axis value being set.