Fix bug in Keyboard where non-alphanumeric keys would fail.
Before, events would look like `keydown:a` and `keydown:;`. With a key `a` this would work, but with `;` it failed. (Slick would return no pseudos.)
Now Keyboard uses the same `keydown:keys(a)` syntax that Element.Pseudos.Keys uses as well.

Also changed the Keyboard specs. The previous specs were for the implementation reverted in commit d036afe.