Changing the drop logic for keyboard.
Previously because drop called Keyboard.manager.manage(instance) it meant that the Keyboard.manager could never drop a keyboard. In theory this is fine, but in effect it meant that despite the fact that keyboards would have other managers assigned, they'd still be in the "instances" array of the Keyboard.manager forever. Now an instance can be dropped, effectively disabling it.