Configure undraggable tags in Sortables.
The input, button, a, textarea, select and option tags aren't draggable
when using Sortables. This patch adds an unDraggableTags option that can
be passed to Sortables to override the above list, allowing you to drag
a container element even when you click on one of the (undraggable)
elements within it.