Fix getComputedSize() in Element.Measure.js
If the style is some string and `this.getStyle(style).toInt()` returns `NaN` then the results value of `totalWidth` and `totalHeight` also will be `NaN`. I recommend to set the style value in those situations to `0`.

I had this problem in Opera and IE. For example, `border-top-width` was automatic set by browser to `medium`. The conversion of this value into integer returns `NaN`. Therefore, after next calculation, `totalWidth` and `totalHeght` will be `NaN`.