Fixes scrolling during drag operations
This pull request fixes the behavior of Drag.
As it is possible to scroll the offsetParent container for example via Mouse-Wheel during a Drag operation,
this scroll offset needs to be accounted for.

You can see the problematic behavior in this jsfiddle:
http://jsfiddle.net/6V4rG/1/

Just use the Mousewheel while resizing/moving a box.

Warning:
The mousemove behavior might not be consistent accross browsers.
The above (by this commit fixed) issue was present under

- Chromium 35.0.1871.0 (254670)
- FF 27.0.1
- IE 9.0.8112.16421

TODO: For a complete handling, a proper mousemove event should be dispatched from the scroll handler as well.